+++ src/Features/CSharp/Portable/ConvertToRawString/ConvertToRawKind.cs
@@ -1,0 +1,12 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace Microsoft.CodeAnalysis.CSharp.ConvertToRawString;
+
+internal enum ConvertToRawKind
+{
+    SingleLine,
+    MultiLineIndented,
+    MultiLineWithoutLeadingWhitespace,
+}
